Modificaciones en las vistas, repositorios, entidades
relice algunas validaciones a nivel de servidor en las entidades y modifique algunos campos
realice algunos cambios en los servicios para implementar el delete en las vistas
agregue los campos q le faltaban a las vistas Area , Estudiante, Libro, Trabajador, Tarjeta de Prestamo de Estudiantes y Trabajadores

+++ src/main/java/com/example/application/data/service/UserService.java
@@ -35,8 +35,8 @@ public class UserService {
         return repository.findAll(pageable);
     }
 
-    public int count() {
-        return (int) repository.count();
+    public long count() {
+        return  repository.count();
     }
     
 }
